Refactor LibraryCommander
- pulled out a bunch of methods dealing with invites into
LibraryInviteCommander (new class)
- modified existing tests for libraryCommander to call relevant methods
from libraryInviteCommander instead.
- updated all previous references of refactored methods to the new
class.
